+++ ledger/daml-on-sql/src/main/scala/on/sql/Cli.scala
@@ -3,7 +3,7 @@
 
 package com.daml.on.sql
 
-import com.daml.ledger.participant.state.v1.SeedService.Seeding
+import com.daml.platform.apiserver.SeedService.Seeding
 import com.daml.platform.common.LedgerIdMode
 import com.daml.platform.sandbox.cli.{CommonCli, SandboxCli}
 import com.daml.platform.sandbox.config.{PostgresStartupMode, SandboxConfig}
