+++ ledger/participant-state/BUILD.bazel
@@ -34,7 +34,6 @@ da_scala_library(
         "//ledger/participant-state/protobuf:ledger_configuration_proto_java",
         "//libs-scala/logging-entries",
         "@maven//:com_google_protobuf_protobuf_java",
-        "@maven//:org_slf4j_slf4j_api",
     ],
 )
