+++ ledger/sandbox-common/src/test/lib/scala/platform/sandbox/AbstractSandboxFixture.scala
@@ -21,9 +21,9 @@ import com.daml.ledger.api.v1.ledger_identity_service.{
 }
 import com.daml.ledger.api.v1.testing.time_service.TimeServiceGrpc
 import com.daml.ledger.client.services.testing.time.StaticTime
-import com.daml.ledger.participant.state.v1.SeedService.Seeding
 import com.daml.ledger.resources.ResourceOwner
 import com.daml.ledger.test.ModelTestDar
+import com.daml.platform.apiserver.SeedService.Seeding
 import com.daml.platform.common.LedgerIdMode
 import com.daml.platform.sandbox.config.SandboxConfig
 import com.daml.platform.sandbox.services.DbInfo
