+++ ledger/sandbox/src/main/scala/platform/sandboxnext/Cli.scala
@@ -3,7 +3,7 @@
 
 package com.daml.platform.sandboxnext
 
-import com.daml.ledger.participant.state.v1.SeedService.Seeding
+import com.daml.platform.apiserver.SeedService.Seeding
 import com.daml.platform.sandbox.cli.{CommonCli, SandboxCli}
 import com.daml.platform.sandbox.config.SandboxConfig
 import scopt.OptionParser
