+++ ledger/sandbox/src/main/scala/platform/sandboxnext/SeedServiceLogEntryIdAllocator.scala
@@ -6,8 +6,8 @@ package com.daml.platform.sandboxnext
 import java.util.UUID
 
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlLogEntryId
-import com.daml.ledger.participant.state.v1.SeedService
 import com.daml.ledger.validator.LogEntryIdAllocator
+import com.daml.platform.apiserver.SeedService
 import com.google.protobuf.ByteString
 
 final class SeedServiceLogEntryIdAllocator(seedService: SeedService) extends LogEntryIdAllocator {
